Revert "implement password reset flow and update task components"
