--- src/components/dashboardComponent/Chart01.vue
@@ -1,126 +1,0 @@
-<template>
-  <div class="chart-container">
-    <Bar
-      id="monthly-registered-users-chart"
-      :options="chartOptions"
-      :data="chartData"
-    />
-  </div>
-</template>
-
-<script>
-import { Bar } from "vue-chartjs";
-import {
-  Chart as ChartJS,
-  Title,
-  Tooltip,
-  Legend,
-  BarElement,
-  CategoryScale,
-  LinearScale,
-} from "chart.js";
-
-ChartJS.register(
-  Title,
-  Tooltip,
-  Legend,
-  BarElement,
-  CategoryScale,
-  LinearScale
-);
-
-export default {
-  name: "MonthlyRegisteredUsersChart",
-  components: {
-    Bar,
-  },
-  data() {
-    return {
-      chartData: {
-        labels: [
-          "Jan",
-          "Feb",
-          "Mar",
-          "Apr",
-          "May",
-          "Jun",
-          "Jul",
-          "Aug",
-          "Sep",
-          "Oct",
-          "Nov",
-          "Dec",
-        ],
-        datasets: [
-          {
-            label: "Monthly Registered Users",
-            backgroundColor: "#3b82f6",
-            data: [17, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-          },
-        ],
-      },
-      chartOptions: {
-        responsive: true,
-        maintainAspectRatio: false,
-        plugins: {
-          title: {
-            display: true,
-            text: "Monthly Registered Users in 2025",
-            align: "start",
-          },
-          tooltip: {
-            callbacks: {
-              label: function (context) {
-                let label = context.dataset.label || "";
-                if (label) {
-                  label += ": ";
-                }
-                if (context.parsed.y !== null) {
-                  label += context.parsed.y;
-                }
-                return label;
-              },
-            },
-          },
-          legend: {
-            display: true,
-          },
-        },
-        scales: {
-          x: {
-            title: {
-              display: true,
-              text: "Month",
-            },
-          },
-          y: {
-            title: {
-              display: true,
-              text: "Value",
-            },
-            beginAtZero: true,
-            max: 18,
-          },
-        },
-      },
-    };
-  },
-  methods: {
-    updateChartData(newData) {
-      this.chartData.datasets[0].data = newData;
-    },
-  },
-};
-</script>
-
-<style scoped>
-.chart-container {
-  background-color: #ffffff;
-  border-radius: 8px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  padding: 20px;
-  width: 80%;
-  margin: auto;
-  height: 400px;
-}
-</style>
