+++ src/components/linksfiles.vue/linke.vue
@@ -3,31 +3,17 @@
     class="pt-4 d-flex gap-3 container-fluid justify-content-center align-items-center bg-light border-1 border-bottom"
   >
     <router-link class="btn btn-warning" to="/" role="button">Home</router-link>
-
+    
     <router-link class="btn btn-success" to="/notice" role="button"
       >notice</router-link
     >
-
-    <router-link class="btn btn-primary" to="/rauf" role="button"
-      >Rauf Task</router-link
-    >
-    <router-link class="btn btn-secondary" to="/layo" role="button"
-      >Layo Task</router-link
-    >
-    <router-link class="btn btn-danger" to="/seun" role="button"
-      >Seun Task</router-link
-    >
-    <router-link class="btn btn-dark" to="/kcee" role="button"
-      >Kcee Task</router-link
-    >
-    <!-- <router-link class="btn btn-info" to="/seyi" role="button">Seyi Task</router-link> -->
 
     <router-link class="btn btn-primary" to="/rauf" role="button">Rauf Task</router-link>
     <router-link class="btn btn-secondary" to="/layo" role="button">Layo Task</router-link>
     <router-link class="btn btn-danger" to="/seun" role="button">Seun Task</router-link>
     <router-link class="btn btn-dark" to="/kcee" role="button">Kcee Task</router-link>
     <router-link class="btn btn-info" to="/seyi" role="button">Seyi Task</router-link>
-
+    <router-link class="btn btn-info" to="/rauf2" role="button"></router-link>
 
     <!-- <router-link
       class="btn btn-success"
