+++ src/router/index.js
@@ -9,7 +9,6 @@ import TaskStepIndicator from "../components/TaskFolder/TaskStepIndicator.vue";
 import notice from "../views/notice.vue";
 import Login from "../views/Login.vue";
 import Register from "../views/Register.vue";
-import ForgetPassword from "../views/ForgetPassword.vue";
 import RaufTask from "../components/TaskFolder/RaufTask.vue";
 import LayoTask from "../components/TaskFolder/LayoTask.vue";
 import SeunTask from "../components/TaskFolder/SeunTask.vue";
@@ -23,7 +22,7 @@ import Rauf2 from "../components/TaskFolder/Rauf2.vue";
 // import ParentDashboard from "../views/ParentDashboard.vue";
 // import TutorDashboard from "../views/TutorDashboard.vue";
 
-import StudentProfile from "../views/Dashboard/StudentProfile.vue";
+import StudentDashboard from "../components/Dashboard.vue";
 
 const routes = [
   { path: "/", name: "Home", component: Home },
@@ -38,17 +37,6 @@ const routes = [
     path: "/register",
     name: "Register",
     component: Register,
-  },
-  {
-    path: "/forget-password",
-    name: "ForgetPassword",
-    component: ForgetPassword,
-  },
-
-  {
-    path: "/reset-password",
-    name: "ResetPassword",
-    component: () => import("@/views/ResetPassword.vue"),
   },
   {
     path: "/taskAdditionalInfoForm",
@@ -119,9 +107,9 @@ const routes = [
   //   component: AdminDashboard,
   // },
   {
-    path: "/student/profile",
-    name: "StudentProfile",
-    component: StudentProfile,
+    path: "/student/dashboard",
+    name: "StudentDashboard",
+    component: StudentDashboard,
   },
   // {
   //   path: "/parentDashboard",
