+++ src/style.css
@@ -73,12 +73,6 @@ button:focus-visible {
   width: 100%;
 }
 
-.dropdown-item.active,
-.dropdown-item:active {
-  background-color: var(--royal-purple);
-  color: white;
-}
-
 /* #app {
   max-width: 1280px;
   margin: 0 auto;
